// types/studentArchive.types.ts
export type StudentImportWizardStep =
  | "file_select"
  | "file_overview"
  | "id_integration"
  | "update_confirm"
  | "final_confirm"
  | "execute"